Add protection against modifications
